# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sansara API Dacumentation",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://notter.loc/api/v1",
            "description": "Sansara Dev Server"
        },
        {
            "url": "http://localhost-prod/api/v1",
            "description": "Sansara Prod Server"
        }
    ],
    "paths": {
        "/auth/refresh-token": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh access token & refresh token",
                "operationId": "refreshToken",
                "parameters": [
                    {
                        "name": "refresh-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "headers": {
                            "Access-Token": {
                                "$ref": "#/components/headers/Access-Token"
                            },
                            "Refresh-token": {
                                "$ref": "#/components/headers/Refresh-token"
                            },
                            "Token-Type": {
                                "$ref": "#/components/headers/Token-Type"
                            },
                            "Expires-In": {
                                "$ref": "#/components/headers/Expires-In"
                            }
                        }
                    },
                    "403": {
                        "description": "Refresh token invalid"
                    }
                }
            }
        },
        "/auth/sign-in": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Авторизация",
                "operationId": "signIn",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/SignIn"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "headers": {
                            "Access-Token": {
                                "$ref": "#/components/headers/Access-Token"
                            },
                            "Refresh-token": {
                                "$ref": "#/components/headers/Refresh-token"
                            },
                            "Token-Type": {
                                "$ref": "#/components/headers/Token-Type"
                            },
                            "Expires-In": {
                                "$ref": "#/components/headers/Expires-In"
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignInErrorsSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/sign-up": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Регистрация",
                "operationId": "signUp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/SignUp"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "headers": {
                            "Access-Token": {
                                "$ref": "#/components/headers/Access-Token"
                            },
                            "Refresh-token": {
                                "$ref": "#/components/headers/Refresh-token"
                            },
                            "Token-Type": {
                                "$ref": "#/components/headers/Token-Type"
                            },
                            "Expires-In": {
                                "$ref": "#/components/headers/Expires-In"
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignUpErrorsSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/user": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Возвращает текущего авторизованного пользователя",
                "operationId": "user",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserUnauthenticatedSchema"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "PostRequest": {
                "title": "Sign in request",
                "description": "Sign-in request",
                "properties": {
                    "email": {
                        "title": "Email",
                        "description": "Email",
                        "type": "integer",
                        "format": "string",
                        "example": "test@test.ru"
                    },
                    "password": {
                        "title": "Password",
                        "description": "password",
                        "type": "string",
                        "format": "string",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
            "AuthResponse": {
                "title": "Auth Response",
                "properties": {
                    "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/UserSchema"
                    }
                },
                "type": "object"
            },
            "SignInErrorsSchema": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "These credentials do not match our records."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SignUpErrorsSchema": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The email field is required"
                                }
                            },
                            "name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The name field is required."
                                }
                            },
                            "password": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The password field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UserUnauthenticatedSchema": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "SignIn": {
                "title": "Sign in schema",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "test@test.ru"
                    },
                    "password": {
                        "type": "string",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
            "SignUp": {
                "title": "Sign up model",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "email": {
                        "type": "string",
                        "example": "test@test.ru"
                    },
                    "password": {
                        "type": "string",
                        "example": "123456"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
            "UserSchema": {
                "title": "User schema",
                "properties": {
                    "user": {
                        "properties": {
                            "id": {
                                "type": "int",
                                "example": "12"
                            },
                            "name": {
                                "type": "string",
                                "example": "Иван"
                            },
                            "email": {
                                "type": "string",
                                "example": "test@test.ru"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "headers": {
            "Access-Token": {
                "description": "Токен",
                "schema": {
                    "type": "string"
                }
            },
            "Refresh-token": {
                "description": "Токен для обновление access токена",
                "schema": {
                    "type": "string"
                }
            },
            "Token-Type": {
                "description": "Тип токена",
                "schema": {
                    "type": "string"
                }
            },
            "Expires-In": {
                "description": "продолжительность жизни токена",
                "schema": {
                    "type": "int"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        }
    ]
}